stock_screener_test/requirements.txt
# Stock Screener Requirements

# Core dependencies
pandas>=1.5.0
numpy>=1.20.0
PyYAML>=6.0
futu-api>=6.0.0

# Technical analysis
TA-Lib>=0.4.0

# Excel export (optional)
openpyxl>=3.0.0

# Scheduling (for automated runs)
schedule>=1.2.0

# Logging and utilities
python-dateutil>=2.8.0

# Development and testing (optional)
pytest>=7.0.0
jupyter>=1.0.0
